Struct medio::Medium

pub struct Medium<'a> {
    // some fields omitted
}

Entry point for access to medium.com api

Methods

impl<'a> Medium<'a>

fn new(client: &'a Client, token: Option<&'a str>) -> Medium<'a>

creates a new instance of a Medium client token can be acquired here

fn host(client: &'a Client, token: Option<&'a str>, host: &'a str) -> Medium<'a>

creates an instance of a Medium client for a specific host

fn me(&self) -> Result<User>

access to the authentied members user info

fn users(&self, id: &'a str) -> UserRef

access to a user reference